requirements.txt
asgiref==3.8.1
certifi==2024.12.14
charset-normalizer==3.4.1
cloudinary==1.42.1
Django==5.0.7
django-cloudinary-storage==0.3.0
django-cors-headers==4.4.0
django-jazzmin==3.0.0
djangorestframework==3.15.2
gunicorn==23.0.0
idna==3.10
packaging==24.2
pillow==10.4.0
psycopg2==2.9.10
requests==2.32.3
six==1.17.0
sqlparse==0.5.0
tzdata==2025.1
urllib3==2.3.0
